Install latest version of OpenSSL on build
Make sure that the version of OpenSSL installed in the image is
the latest. This ensures that we get the patched version 1.1.0j that
came in late November 2018 and that resolves security bugs.

This shouldn't cause any dependency problems since it's a patch.
Additionally, @jwalgran noted that the dependency tree shows openssl
is only required as a dependency to ca-certificates, and has a wide
version range:

https://github.com/azavea/american-water-tank-management/issues/84#issuecomment-444215952